factor out login wrapper
